# openapi.yaml
openapi: 3.0.3
info:
  title: ms-productos API
  description: |
    Microservicio de gestión del catálogo de productos.
    Permite registrar, listar, buscar, actualizar y eliminar productos.
    La eliminación es lógica (estado → false).
  version: 1.0.0
  contact:
    name: Miguel Sánchez
    email: sanchezsanchezmiguelivan@gmail.com

servers:
  - url: http://localhost:8080
    description: Local
  - url: https://ms-productos.onrender.com
    description: Render (producción)

tags:
  - name: Productos
    description: Operaciones del catálogo de productos

paths:
  /api/productos:
    post:
      tags: [Productos]
      summary: Crear producto
      description: Registra un nuevo producto. La fecha de creación se asigna automáticamente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoRequest'
            example:
              nombre: Laptop Lenovo
              descripcion: Laptop para desarrollo de software
              precio: 3500.00
              stock: 10
              estado: true
      responses:
        '201':
          description: Producto creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoResponse'
              example:
                id: 1
                nombre: Laptop Lenovo
                descripcion: Laptop para desarrollo de software
                precio: 3500.00
                stock: 10
                estado: true
                fechaCreacion: "2026-05-17T15:42:29.832"
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                mensaje: Datos inválidos
                detalle: El nombre no puede estar vacío, El precio debe ser mayor que cero
                fecha: "2026-05-17T15:42:00"

    get:
      tags: [Productos]
      summary: Listar productos
      description: Retorna todos los productos registrados (activos e inactivos).
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductoResponse'

  /api/productos/{id}:
    get:
      tags: [Productos]
      summary: Buscar producto por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoResponse'
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                mensaje: Producto no encontrado
                detalle: No existe un producto con el ID 999
                fecha: "2026-05-17T15:42:46.980"

    put:
      tags: [Productos]
      summary: Actualizar producto
      description: Actualiza todos los campos de un producto existente.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductoRequest'
            example:
              nombre: Laptop Lenovo IdeaPad
              descripcion: Laptop actualizada
              precio: 3200.00
              stock: 8
              estado: true
      responses:
        '200':
          description: Producto actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductoResponse'
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [Productos]
      summary: Eliminar producto (lógico)
      description: Desactiva el producto cambiando el campo `estado` a `false`. No elimina físicamente el registro.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '204':
          description: Producto desactivado (sin cuerpo de respuesta)
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ProductoRequest:
      type: object
      required:
        - nombre
        - precio
        - stock
      properties:
        nombre:
          type: string
          description: Nombre del producto (obligatorio, no vacío)
          example: Laptop Lenovo
        descripcion:
          type: string
          description: Descripción breve del producto (opcional)
          example: Laptop para desarrollo de software
        precio:
          type: number
          format: decimal
          description: Precio unitario (debe ser mayor que cero)
          minimum: 0.01
          example: 3500.00
        stock:
          type: integer
          description: Cantidad disponible (no puede ser negativa)
          minimum: 0
          example: 10
        estado:
          type: boolean
          description: Estado del producto (default true si no se envía)
          example: true

    ProductoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Identificador único generado automáticamente
          example: 1
        nombre:
          type: string
          example: Laptop Lenovo
        descripcion:
          type: string
          example: Laptop para desarrollo de software
        precio:
          type: number
          format: decimal
          example: 3500.00
        stock:
          type: integer
          example: 10
        estado:
          type: boolean
          description: "true = activo, false = inactivo (eliminado lógicamente)"
          example: true
        fechaCreacion:
          type: string
          format: date-time
          description: Asignada automáticamente al crear el registro
          example: "2026-05-17T15:42:29.832"

    ErrorResponse:
      type: object
      properties:
        mensaje:
          type: string
          description: Descripción breve del error
          example: Producto no encontrado
        detalle:
          type: string
          description: Detalle específico del error
          example: No existe un producto con el ID 999
        fecha:
          type: string
          format: date-time
          description: Fecha y hora en que ocurrió el error
          example: "2026-05-17T15:42:46.980"